add ended status tracking for mods
